requirements.txt
asgiref==3.4.1
boto3==1.20.24
botocore==1.23.24
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.8
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.0
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.9
django-cors-headers==3.10.0
django-environ==0.8.1
django-heroku==0.3.1
django-rest-passwordreset==1.2.1
django-storages==1.12.3
django-templated-mail==1.1.1
django-versatileimagefield==2.2
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-flex-fields==0.9.6
gunicorn==20.1.0
idna==3.3
importlib-metadata==1.7.0
itypes==1.2.0
Jinja2==3.0.3
jmespath==0.10.0
libmagic==1.0
MarkupSafe==2.0.1
oauthlib==3.1.1
Pillow==8.4.0
psycopg2==2.9.2
pycparser==2.21
PyJWT==2.3.0
python-dateutil==2.8.2
python-magic-bin==0.4.14
python3-openid==3.2.0
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
s3transfer==0.5.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.2
typing-extensions==4.0.0
uritemplate==4.1.1
urllib3==1.26.7
whitenoise==5.3.0
zipp==3.6.0
